trainer updates
