+++ src/Entities/Trainer.ts
@@ -1,0 +1,9 @@
+interface ITrainer
+{
+        trainerid: number,
+        email: string,
+        trainerfirst: string,
+        trainerlast: string
+}
+
+export default ITrainer;
