--- src/components/trainers/ViewTrainer/ViewTrainer.test.tsx
@@ -1,38 +1,0 @@
-import React from 'react';
-import ViewTrainer from '.';
-import { shallow, mount } from 'enzyme';
-/**
- * Authors: Joab Smith and Imran Ilyas
- **/
-describe('ViewTrainer', () => {
-  const trainer = {
-    FirstName: 'John',
-    LastName: 'Doe',
-    Email: 'johndoe@hotmail.com',
-  };
-
-  const mock = jest.fn();
-  const wrapper = mount(<ViewTrainer trainer={trainer} setEdit={mock} />);
-  const shallowWrapper = shallow(
-    <ViewTrainer trainer={trainer} setEdit={mock} />
-  );
-
-  it('Should have all text fields', () => {
-    expect(shallowWrapper.find('Text')).toHaveLength(8);
-  });
-
-  it('should have an edit button', () => {
-    expect(wrapper.find('TouchableOpacity')).toBeDefined();
-  });
-  it('should call Edit function when button is pressed', () => {
-    const sub = wrapper
-      .find('TouchableOpacity')
-      .findWhere((w) => w.text() === 'Edit')
-      .first();
-    const mockEventHandler = jest.spyOn(sub.props(), 'onPress');
-    // Enzyme usually allows wrapper.simulate() alternatively, but this doesn't support 'press' events.
-    sub.props().onPress();
-    expect(mockEventHandler).toHaveBeenCalled();
-    expect(mock).toHaveBeenCalledWith(true);
-  });
-});
